requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pymongo==4.14.1
python-dotenv==1.1.1
sqlparse==0.5.3
